Forbid multiple assignment targets for context managers #328 Review
