+++ tests/test_checker/test_noqa.py
@@ -61,7 +61,6 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z222': 1,
         'Z223': 1,
         'Z224': 1,
-        'Z225': 0,
 
         'Z300': 1,
         'Z301': 1,
@@ -79,6 +78,7 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z313': 1,
         'Z314': 1,
         'Z315': 1,
+        'Z316': 0,
 
         'Z400': 0,
         'Z401': 0,
