+++ tests/test_visitors/test_ast/test_complexity/test_counts/test_context_manager_targets.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from wemake_python_styleguide.violations.complexity import (
+from wemake_python_styleguide.violations.consistency import (
     TooManyContextManagerAssignmentsViolation,
 )
 from wemake_python_styleguide.visitors.ast.complexity.counts import (
@@ -14,9 +14,20 @@ with open('') as first, second:
     ...
 """
 
+correct_count = """
+with open('') as first:
+    ...
+"""
+
+composite_assign = """
+with open('first') as first, open('second') as second:
+    ...
+"""
+
 
 @pytest.mark.parametrize('code', [
     incorrect_count,
+    composite_assign,
 ])
 def test_context_manager_multiple_targets(
     assert_errors,
@@ -33,23 +44,13 @@ def test_context_manager_multiple_targets(
     assert_errors(visitor, [TooManyContextManagerAssignmentsViolation])
 
 
-correct_count = """
-with open('') as first:
-    ...
-"""
-
-
-@pytest.mark.parametrize('code', [
-    correct_count,
-])
 def test_context_manager_alone_target(
     assert_errors,
     parse_ast_tree,
-    code,
     default_options,
 ):
     """Testing correct count context manager assignment."""
-    tree = parse_ast_tree(code)
+    tree = parse_ast_tree(correct_count)
 
     visitor = ContextManagerVisitor(default_options, tree=tree)
     visitor.run()
