+++ wemake_python_styleguide/violations/consistency.py
@@ -41,6 +41,7 @@ Summary
    MissingSpaceBetweenKeywordAndParenViolation
    WrongConditionalViolation
    ObjectInBaseClassesListViolation
+   TooManyContextManagerAssignmentsViolation
 
 Consistency checks
 ------------------
@@ -61,6 +62,7 @@ Consistency checks
 .. autoclass:: MissingSpaceBetweenKeywordAndParenViolation
 .. autoclass:: WrongConditionalViolation
 .. autoclass:: ObjectInBaseClassesListViolation
+.. autoclass:: TooManyContextManagerAssignmentsViolation
 
 """
 
@@ -581,3 +583,31 @@ class ObjectInBaseClassesListViolation(ASTViolation):
     should_use_text = False
     error_template = 'Founded extra `object` in parent classes list'
     code = 315
+
+
+@final
+class TooManyContextManagerAssignmentsViolation(ASTViolation):
+    """
+    Forbid multiple assignment targets for context managers.
+
+    Reasoning:
+        Following "as" with another context manager looks like a tuple.
+        Emitted when a with statement component returns multiple values
+        and uses name binding with as only for a part of those values,
+        as in with ctx() as a, b.This can be misleading, since it's not
+        clear if the context manager returns a tupleor if the node
+        without a name binding is another context manager.
+
+    Example::
+
+        # Wrong:
+        with open('') as first, second:
+            ...
+
+    .. versionadded:: 0.6.0
+
+    """
+
+    should_use_text = False
+    error_template = 'Found context manager with too many assignments'
+    code = 316
